Added collision area code
Added code to load the collision area data from the course JSON, and to detect whether the player car is inside a collision area

Added code to respond to the three types of offroad (light, medium, heavy) and the shake area type

Added CSS code to have the game canvas fill all available space on the screen

Added background sprites data for the offroad and out of bounds on the Circuit track

Added code to activate and deactivate debug mode (shows the collision area overlay in debug mode and prints some messages in the browser console)

Added code to deserialize geometry shapes and collision areas from JSON

Added code to use a game loop that keeps track of the time elapsed between the loop iterations, and added code to PhysicsObject and Car to use time delta based movement

Added code to GameState to have a singleton state object for the game

Added code to Car to slow down the car in the offroad, shake the camera in the offroad and CAMERA_SHAKE collision area types, and reduce the turning radius of the car in the offroad collision areas. Also changed the car so that it can rotate slightly on screen in offroad areas, to show that it is skidding. Also changed the angle code to use angular velocity and time delta based angle updates.