Added main menu, choose car menu, and pause menu
Added HTML and CSS code for the main menu, and the waving_flag.svg logo displayed there
Added Press Start 2P font used for the text on the main menu and the debug screen

Added Choose Car menu to allow the player to choose between the red, green, blue and purple cars

Added pause menu so that the player can pause the game by pressing the Escape key while driving the car, and all rendering and physics object updates will be immediately stopped. Also added code to communicate between the game's worker thread and the main thread regarding the pausing of the game. Also added code to Keyboard to prevent holding down the ESC key from repeatedly triggering the pause/unpause event.

Added debug text to debug mode, which shows the position, velocity and angle of the player car